requirements.txt
py-solc-x==1.1.1
web3==5.4.0